requirements.txt
absl==0.0
absl_py==0.15.0
anyascii==0.3.1
coqpit==0.0.17
editdistance==0.6.1
Flask==1.1.2
fsspec==2022.3.0
g2pkk==0.1.2
gruut==2.3.4
gruut_ipa==0.13.0
inflect==5.3.0
jamo==0.4.1
jieba==0.42.1
kaggle==1.5.12
kaldiio==2.17.2
librosa==0.9.2
matplotlib==3.5.1
MeCab==0.996.3
num2words==0.5.12
numpy==1.19.5
pandas==1.1.5
pyaudio==0.2.13
pygame==2.1.2
pypinyin==0.47.1
pysbd==0.3.4
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.0
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
tensorflow==2.11.0
tensorflow_gpu==2.6.0
textgrid==1.5
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
tqdm==4.62.3
trainer==0.0.20
umap==0.1.1
umap_learn==0.5.2
Unidecode==1.3.2
visdom==0.1.8.9
